Relocate layout helpers to utils
